ValueConverter - Adds null reference check
If a MegaNav value is emptied out in the CMS, then the `source` value is an empty string.
The empty string is then attempted to be deserialized to a list of MeganavItem items - which would be `null`.
Then the null-reference is passed to the `BuildMenu` method, which then `ToList()` is immediately called - throwing an exception.

This exception is caught and logged, but it is preventable by adding a null/empty-string check upfront - saves the extra work.